Split real-time-logger.h into two files
Put class declarations in one and macro definitions in the other.
